requirements.txt
torch==2.4.0
torchvision==0.19.0
numpy~=1.24.4
scipy==1.9.1
pandas==2.0.3
pytorch-msssim==1.0.0
numpy==1.24.3
tqdm
matplotlib~=3.7.5
pillow~=10.4.0